[refactor] feature bookmark, search 추가
