+++ src/features/bookmark/index.ts
@@ -1,3 +1,5 @@
 export { default as BookmarkButton } from './ui/BookmarkButton';
 export { default as BookmarkItem } from './ui/BookmarkItem';
 export { default as BookmarkList } from './ui/BookmarkList';
+
+export { useBookmarks, useBookmarkActions } from './model/useBookmarkStore';
