+++ src/features/forecast/model/useForecastQuery.ts
@@ -1,10 +1,10 @@
 import { useQuery } from '@tanstack/react-query';
 
-import ForecastService from '@/features/forecast/api/ForecastAPI';
+import ForecastAPI from '../api/ForecastAPI';
 import {
   getShortTermForecastBaseDateTime,
   getUltraShortTermForecastBaseDateTime,
-} from '@/features/forecast/utils/getForecastBaseDateTime';
+} from '../utils/getForecastBaseDateTime';
 
 export const useFetchUltraShortTermForecast = (params: {
   x: number;
@@ -16,7 +16,7 @@ export const useFetchUltraShortTermForecast = (params: {
   return useQuery({
     queryKey: ['forecast', { type: 'ultra' }, params.x, params.y],
     queryFn: () =>
-      ForecastService.getUltraShortTermForecast({ baseDate, baseTime, x, y }),
+      ForecastAPI.getUltraShortTermForecast({ baseDate, baseTime, x, y }),
     enabled: x !== 0 && y !== 0,
     staleTime: 1000 * 10, // 10분 이내에는 캐시된 결과를 사용
   });
@@ -29,7 +29,7 @@ export const useFetchShortTermForecast = (params: { x: number; y: number }) => {
   return useQuery({
     queryKey: ['forecast', { type: 'short' }, params.x, params.y],
     queryFn: () =>
-      ForecastService.getShortTermForecast({ baseDate, baseTime, x, y }),
+      ForecastAPI.getShortTermForecast({ baseDate, baseTime, x, y }),
     enabled: x !== 0 && y !== 0,
     staleTime: 1000 * 10,
   });
