+++ src/features/search/model/types.ts
